Fix bugs here and there, added pay command, use Int for balance, closes #14

+++ src/main/kotlin/dev/skrub/thunderhead/dataclass/economy/Human.kt
@@ -2,6 +2,6 @@ package dev.skrub.thunderhead.dataclass.economy
 
 data class Human(
     val id: Long,
-    var balance: Long,
+    var balance: Int,
     var lastDaily: Long
 )
